requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
accelerate>=0.24.0
datasets>=2.14.0
Pillow>=10.0.0
numpy>=1.24.0
scipy>=1.10.0
einops>=0.7.0
timm>=0.9.0
safetensors>=0.4.0
kaggle>=1.5.0
opencv-python>=4.8.0
matplotlib>=3.7.0
seaborn>=0.12.0
wandb>=0.15.0
tqdm>=4.65.0
psutil>=5.9.0
memory-profiler>=0.61.0
bitarray>=2.8.0
numba>=0.58.0
cython>=3.0.0
pybind11>=2.11.0

